feat: implement file upload and deletion functionality, refactor chat controller for improved error handling
